service/logs: move to cli/internal/logdetails
This package is only used by cli/command/service, and has no
external consumers.
