+++ cli/command/service/logs.go
@@ -13,7 +13,7 @@ import (
 	"github.com/docker/cli/cli/command"
 	"github.com/docker/cli/cli/command/completion"
 	"github.com/docker/cli/cli/command/idresolver"
-	"github.com/docker/cli/service/logs"
+	"github.com/docker/cli/cli/internal/logdetails"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/swarm"
@@ -267,7 +267,7 @@ func (lw *logWriter) Write(buf []byte) (int, error) {
 		return 0, errors.Errorf("invalid context in log message: %v", string(buf))
 	}
 	// parse the details out
-	details, err := logs.ParseLogDetails(string(parts[detailsIndex]))
+	details, err := logdetails.Parse(string(parts[detailsIndex]))
 	if err != nil {
 		return 0, err
 	}
